+++ backend/apid/actions/entities_test.go
@@ -493,3 +493,99 @@ func TestEntityCreate(t *testing.T) {
 		})
 	}
 }
+
+func TestEntityCreateOrReplace(t *testing.T) {
+	defaultCtx := testutil.NewContext(
+		testutil.ContextWithOrgEnv("default", "default"),
+		testutil.ContextWithRules(
+			types.FixtureRuleWithPerms(
+				types.RuleTypeEntity,
+				types.RulePermCreate,
+				types.RulePermUpdate,
+			),
+		),
+	)
+	wrongPermsCtx := testutil.NewContext(
+		testutil.ContextWithOrgEnv("default", "default"),
+		testutil.ContextWithRules(
+			types.FixtureRuleWithPerms(types.RuleTypeEntity, types.RulePermRead),
+		),
+	)
+
+	badEntity := types.FixtureEntity("badentity")
+	badEntity.ID = ""
+
+	testCases := []struct {
+		name            string
+		ctx             context.Context
+		argument        *types.Entity
+		fetchResult     *types.Entity
+		fetchErr        error
+		createErr       error
+		expectedErr     bool
+		expectedErrCode ErrCode
+	}{
+		{
+			name:        "Created",
+			ctx:         defaultCtx,
+			argument:    types.FixtureEntity("foo"),
+			fetchResult: nil,
+			fetchErr:    nil,
+			createErr:   nil,
+			expectedErr: false,
+		},
+		{
+			name:        "Already exists",
+			ctx:         defaultCtx,
+			argument:    types.FixtureEntity("foo"),
+			fetchResult: types.FixtureEntity("foo"),
+		},
+		{
+			name:            "No permission",
+			ctx:             wrongPermsCtx,
+			argument:        types.FixtureEntity("foo"),
+			fetchResult:     nil,
+			expectedErr:     true,
+			expectedErrCode: PermissionDenied,
+		},
+		{
+			name:            "Validation error",
+			ctx:             defaultCtx,
+			argument:        badEntity,
+			expectedErr:     true,
+			expectedErrCode: InvalidArgument,
+		},
+	}
+
+	for _, tc := range testCases {
+		store := &mockstore.MockStore{}
+		actions := NewEntityController(store)
+
+		t.Run(tc.name, func(t *testing.T) {
+			assert := assert.New(t)
+
+			// Mock store methods
+			store.
+				On("GetEntityByID", mock.Anything, mock.Anything).
+				Return(tc.fetchResult, tc.fetchErr)
+			store.
+				On("UpdateEntity", mock.Anything, mock.Anything).
+				Return(tc.createErr)
+
+			// Exec Query
+			err := actions.CreateOrReplace(tc.ctx, *tc.argument)
+
+			if tc.expectedErr {
+				inferErr, ok := err.(Error)
+				if ok {
+					assert.Equal(tc.expectedErrCode, inferErr.Code)
+				} else {
+					assert.Error(err)
+					assert.FailNow("Given was not of type 'Error'")
+				}
+			} else {
+				assert.NoError(err)
+			}
+		})
+	}
+}