+++ cli/commands/entity/help.go
@@ -14,6 +14,7 @@ func HelpCommand(cli *cli.SensuCli) *cobra.Command {
 
 	// Add sub-commands
 	cmd.AddCommand(
+		CreateCommand(cli),
 		DeleteCommand(cli),
 		ListCommand(cli),
 		InfoCommand(cli),