+++ cli/commands/helpers/flags.go
@@ -16,7 +16,7 @@ var commaWhitespaceRegex *regexp.Regexp
 // configuration the user's configured default format is used as the flag's
 // default value.
 func AddFormatFlag(flagSet *pflag.FlagSet) {
-	flagSet.String("format", config.DefaultFormat, `format of data returned ("json"|"tabular")`)
+	flagSet.String("format", config.DefaultFormat, `format of data returned ("json"|config.FormatTabular)`)
 }
 
 // AddAllOrganization adds the '--all-organizations' flag to the given command