+++ cli/commands/mutator/info.go
@@ -2,6 +2,7 @@ package mutator
 
 import (
 	"errors"
+	"fmt"
 	"io"
 	"strconv"
 
@@ -19,8 +20,6 @@ func InfoCommand(cli *cli.SensuCli) *cobra.Command {
 		Short:        "show detailed mutator information",
 		SilenceUsage: true,
 		RunE: func(cmd *cobra.Command, args []string) error {
-			format, _ := cmd.Flags().GetString("format")
-
 			if len(args) != 1 {
 				_ = cmd.Help()
 				return errors.New("invalid argument(s) received")
@@ -33,11 +32,10 @@ func InfoCommand(cli *cli.SensuCli) *cobra.Command {
 				return err
 			}
 
-			if format == "json" {
-				return helpers.PrintJSON(r, cmd.OutOrStdout())
-			}
-			printToList(r, cmd.OutOrStdout())
-			return nil
+			// Determine the format to use to output the data
+			flag := helpers.GetChangedStringValueFlag("format", cmd.Flags())
+			format := cli.Config.Format()
+			return helpers.PrintFormatted(flag, format, r, cmd.OutOrStdout(), printToList)
 		},
 	}
 
@@ -46,7 +44,11 @@ func InfoCommand(cli *cli.SensuCli) *cobra.Command {
 	return cmd
 }
 
-func printToList(mutator *types.Mutator, writer io.Writer) {
+func printToList(v interface{}, writer io.Writer) error {
+	mutator, ok := v.(*types.Mutator)
+	if !ok {
+		return fmt.Errorf("%t is not a Mutator", v)
+	}
 	cfg := &list.Config{
 		Title: mutator.Name,
 		Rows: []*list.Row{
@@ -73,5 +75,5 @@ func printToList(mutator *types.Mutator, writer io.Writer) {
 		},
 	}
 
-	list.Print(writer, cfg)
+	return list.Print(writer, cfg)
 }